fastapi-server/requirements.txt
fastapi
uvicorn
requests 
python-dotenv
httpx
slowapi
python-multipart
gTTS
